React Marketplace Changes Added FdK CLI

+++ src/lib/Auth.ts
@@ -13,21 +13,20 @@ function getLocalBaseUrl(isTesting = false) {
     return `http${isTesting ? '' : 's'}://localhost`;
 }
 async function checkTokenExpired(auth_token) {
-    const { expiry_time } = auth_token
+    const { expiry_time } = auth_token;
     const currentTimestamp = Math.floor(Date.now() / 1000);
     if (currentTimestamp > expiry_time) {
         return true;
-    }
-    else{
-        return false
+    } else {
+        return false;
     }
 }
 
-export const getApp = async ({isTesting = false}) => {
+export const getApp = async ({ isTesting = false }) => {
     const app = require('https-localhost')(getLocalBaseUrl(isTesting));
     let certs = null;
 
-    if(!isTesting){
+    if (!isTesting) {
         certs = await app.getCerts();
     }
 
@@ -36,8 +35,7 @@ export const getApp = async ({isTesting = false}) => {
 
     app.post('/token', async (req, res) => {
         try {
-            if(Auth.isOrganizationChange)
-                ConfigStore.delete(CONFIG_KEYS.AUTH_TOKEN);
+            if (Auth.isOrganizationChange) ConfigStore.delete(CONFIG_KEYS.AUTH_TOKEN);
             const expiryTimestamp = Math.floor(Date.now() / 1000) + req.body.auth_token.expires_in;
             req.body.auth_token.expiry_time = expiryTimestamp;
             ConfigStore.set(CONFIG_KEYS.AUTH_TOKEN, req.body.auth_token);
@@ -54,11 +52,13 @@ export const getApp = async ({isTesting = false}) => {
     return { app, certs };
 };
 
-export const startServer = async ({isTesting = false}) => {
-    if (Auth.server) return Auth.server
+export const startServer = async ({ isTesting = false }) => {
+    if (Auth.server) return Auth.server;
 
-    const { app, certs } = await getApp({isTesting});
-    const serverIn = isTesting ? require('http').createServer(app) : require('https').createServer(certs, app);
+    const { app, certs } = await getApp({ isTesting });
+    const serverIn = isTesting
+        ? require('http').createServer(app)
+        : require('https').createServer(certs, app);
     Auth.server = serverIn.listen(port, err => {
         if (err) console.log(err);
     });
@@ -96,13 +96,12 @@ export default class Auth {
         const env = ConfigStore.get(CONFIG_KEYS.CURRENT_ENV_VALUE);
         try {
             let domain = null;
-            if(AVAILABLE_ENVS[env] ){
-                let partnerDomain = AVAILABLE_ENVS[env].replace("api.", "partners.")
+            if (AVAILABLE_ENVS[env]) {
+                let partnerDomain = AVAILABLE_ENVS[env].replace('api', 'partners');
                 domain = `https://${partnerDomain}`;
-            }
-            else {
-                let partnerDomain = env.replace("api.", "partners.")
-                domain =`https://${partnerDomain}`
+            } else {
+                let partnerDomain = env.replace('api', 'partners');
+                domain = `https://${partnerDomain}`;
             }
             await open(
                 `${domain}/organizations/?fdk-cli=true&callback=${getLocalBaseUrl()}:${port}`
@@ -144,14 +143,11 @@ export default class Auth {
     }
     private static isAlreadyLoggedIn = async () => {
         const auth_token = ConfigStore.get(CONFIG_KEYS.AUTH_TOKEN);
-        if (auth_token && auth_token.access_token){
-            const isTokenExpired = await checkTokenExpired(auth_token)
-            if(!isTokenExpired)
-                return true;
-            else 
-                return false;
-        }
-        else return false;
+        if (auth_token && auth_token.access_token) {
+            const isTokenExpired = await checkTokenExpired(auth_token);
+            if (!isTokenExpired) return true;
+            else return false;
+        } else return false;
     };
     static stopSever = async () => {
         Auth.server.close(() => {});
